+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   Ultimately this should make no difference since the Nix store will reset all
   ownership permissions anyway, but it may avoid some spurious errors on certain
   systems.
+* `buildPackage` and `buildDepsOnly` now has a `cargoBuildExtraArgs` option for
+  passing arguments specifically to `cargoBuildCommand`.
 
 ## [0.23.2] - 2026-03-23